+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Duplicated signature showing with click on button (better for mobile)
 - Tweaked action buttons style in `struct` view to work better for touch devices
 - Changed `struct` view to open a link in a new tab
+- Added `limit` and `collapsedLimit` options support for `struct` view
 
 ## 1.0.0-beta.20 (09-09-2019)